Move serveStaticFile into the remote runtime
For a more consistent API.
