+++ packages/pwa-kit-runtime/src/ssr/server/express.js
@@ -415,29 +415,6 @@ export const getResponseFromCache = ({req, res, namespace, key}) => {
 }
 
 /**
- * Serve static files from the app's build directory and set default
- * cache-control headers.
- * @since v2.1.0
- *
- * This is a wrapper around the Express `res.sendFile` method.
- *
- * @param {String} filePath - the location of the static file relative to the build directory
- * @param {Object} opts - the options object to pass to the original `sendFile` method
- */
-export const serveStaticFile = (filePath, opts = {}) => {
-    return (req, res) => {
-        const options = req.app.options
-        const file = path.resolve(options.buildDir, filePath)
-        res.sendFile(file, {
-            headers: {
-                [CACHE_CONTROL]: options.defaultCacheControl
-            },
-            ...opts
-        })
-    }
-}
-
-/**
  * Provided for use by requestHook overrides.
  *
  * Call this to return a res that is a redirect to a bundle asset.
