+++ packages/template-retail-react-app/app/ssr.js
@@ -7,7 +7,7 @@
 'use strict'
 
 const path = require('path')
-const {getRuntime, serveStaticFile} = require('pwa-kit-runtime/ssr/server/express')
+const {getRuntime} = require('pwa-kit-runtime/ssr/server/express')
 const {isRemote} = require('pwa-kit-runtime/utils/ssr-server')
 const {getConfig} = require('pwa-kit-runtime/utils/ssr-config')
 const helmet = require('helmet')
@@ -61,7 +61,7 @@ const {handler} = runtime.createHandler(options, (app) => {
     app.get('/callback?*', (req, res) => {
         res.send()
     })
-    app.get('/robots.txt', serveStaticFile('static/robots.txt'))
+    app.get('/robots.txt', runtime.serveStaticFile('static/robots.txt'))
 
     app.get('/worker.js(.map)?', runtime.serveServiceWorker)
     app.get('*', runtime.render)
